Fix for Bug [770838] SIMON1DOS: Fatal crashes with French version
Fix variable underflow for large subtitle strings with no
discernable whitespace. This fixes crashes for French subtitles
that do, in fact, have whitespace but perhaps not using the
same ASCII character that we scan for. Anyway, the text will
display, albeit with an extra blank line, and ScummVM no longer
blows up.

svn-id: r9009