Fix for Bug [775534] 0.5.0 RC SIMON1DOS Random sfx notes during music
Corrected a problem with resource size computation for GMF sound
effect resources. Since GMF resources have no size info, we were
computing sizes with the assumption that each GMF appears in its
own file, so the file size becomes the resource size. This is true
for simon1dos music files but not for SFX files. This fix uses the
resource offset pointers at the beginning of SFX files to properly
compute the size of SFX resources.

svn-id: r9155