Fix for Bug [780918] SAM: Music stops at Bumpusville
Tweaked handling of ImSetTrigger when the triggered
command starts a new song. If the song to be started
is already playing, we stop it (in preparation for
a restart) ONLY if the song containing the command
trigger is actually playing. Otherwise the restart
is likely to never occur (as is the case with the
Bumpusville mansion interior music).

svn-id: r9576