Expanded iMuse debugger commands. All iMuse commands are now under "imuse", and underscores have been removed.
imuse panic - Stop all music tracks
imuse multimidi on/off - Set multimidi
imuse play # - Play a sound resource
imuse play random - Play a random sound resource
imuse stop # - Stop a music resource
imuse stop all - Stop all music resources

Note that "play #" and "play random" only attempt
to play a sound resource. If the resource is not
a music resource or is in a format not supported by
the current music driver, it won't play.

You can use this as a mini jukebox, but be sure
you're in a room with minimal activity. The
goal is to minimize interference from the scripts
while playing sounds that are not expected in
that room.

svn-id: r9641