Allowed derefActor() and derefActorSafe() to treat actor 0 as a valid actor since the original interpreters appear to have allowed it. This fixes bug #786380 ("FOA - FISTS, script crash in Crete").
I do not know what the purpose of actor 0 was in the original. Either it's
a no-op, or it's used to store actor default values.

Note that in the FOA case the room-33-200 script will loop - at least with
the provided savegame - until you leave the room. However, this is quite
harmless.

svn-id: r9650