FadeServer() is now called from ServiceWindows(), thus eliminating the need for the making it a timer handler. This should eliminate the occasional glitches I've been seeing with fades not being completed.
I'm also hoping that it will fix the problem where the game would sometimes
hang when moving between rooms. I know that at least once when I had that
happen to me the game was busy-waiting for the palette to fade.

At the very least, it's one place less to worry about thread-safety in.

svn-id: r9854