Feature Request [806641] Increase MIDI performance by omitting redundancies
Removed the use of the _bank member for iMuse Parts.
The member is still being saved/loaded, and a
warning will display if a non-zero _bank value is
set. If it turns out some musical passage is actually
using the _bank, we'll need to reinstate it.
Otherwise, after sufficient testing we can obsolete
it in the savegames.

svn-id: r10261