Added EXECUTABLE to the MinGW makefile.
This slows down the build procedure enormously, but on the other hand it
actually does something now. :-)

Though, for some reason, make always crashes shortly before the end when I
build it from scratch. I have to repeat the command to get it to finish.
Time to check if there's a newer MinGW version to download, I guess...

svn-id: r10293