The usual cleanup. Removed the "run from CD" check.
This pretty much concludes the first stage of the engine cleanup. All of
the files, except for console/debugging stuff and possibly some header
files, have been changed to use the ScummVM brace style.

As for the console, that one could probably do with some rewriting, in
which case cleaning it up first would just be unnecessary work.

The next stages of the cleanup should include renaming of variables and
functions to follow the ScummVM coding standards, and turning everything
into C++ classes. And so on.

Of course, the driver directory should go through a similar cleanup as
well.

This has all been enormously tedious, so don't count on me doing any of
these things at the moment. Particularly not turning everything into C++
classes. I'm really not that familiar with C++. :-)

svn-id: r10340