Some renaming (ironic in the light of Fingolfin's recent namespace suggestion, but I prepared the patch long before reading the mail :-).
Also, the remaining parts of the control panel etc. have been moved into a
class of their own.

This is still work in progress. I'm well aware that some of the classes
aren't as well separated as they ought to be, and that using global
variables to keep track of the different classes probably isn't pretty.

svn-id: r10561