Make sure that playSample() gets a pointer to the original QueueElement instead of a copy of it. Otherwise the sound engine will never notice when a sample finishes playing since it's looking at the wrong sound handle.
For whatever reason, this also seems to fix the "game crashes occasionally
on startup with completely useless stack trace" bug for me. The crash was
easily repeatable in Valgrind for me, but the messages it produced were
just as unhelpful as the stack trace. These messages are also gone now.

svn-id: r11757