This could be wrong so feel free to revert, but it does seem to fix crashes in the Mine Road sequence of Full Throttle for me.
The crashes were always in proc64(), with 'en' being some ridiculously
large value. The function, as it was written, did not make any sense to me
so I rewrote it with the assumption that it's there to remove the "en"th
element of the _val216d[] (remaining enemies?) array.

I don't know how well it will work from proc63() if there's more than one
element to remove - at least if the elements to remove are neighbours - but
maybe that case never happens.

svn-id: r11999