Set up file paths to handle the differences between the file layout of the Windows and Linux versions. (TODO: Handle game detection.)
Added preliminary support for the P2_A.aif file that the Linux version
uses, where the Windows version uses p2_a.voc. Someone will have to verify
the actual format of it, and we have a silly bug in playVoice() which will
have to be fixed before things work properly.

svn-id: r15365