Added music to the partial IHNM intro. I think I picked the correct tune, but I can't be absolutely sure.
I'm also not sure if I'm using the correct music data. I'm fairly sure that
the MUSICGM.RES file is the appropriate choice for "General MIDI", but I
have no way of verifying that the MUSICFM.RES is really the right choice
for MT-32. (Is either of them a good choice for Adlib, or are they both
equally bad?)

svn-id: r15831