After listening to some of the IHNM music from both the FM and GM music files through the MIDI synth on my Soundblaster, I can only conclude that they are both General MIDI, so I've changed our code to assume that in both cases. The End Title (which we currently incorrectly play at the end of the intro) sounds much better in Adlib this way, too.
Perhaps the original used a different GM to FM mapping than the one in our
adlib.cpp, but if so using the MT-32 to GM mapping isn't likely to improve
the situation.

We could add a way of replacing the GM to FM mapping in adlib.cpp, I guess.
Of course, that would only be worth it if the original sounds better than
our mapping.

svn-id: r15894