Fixed an Adlib regression with older games (e.g. EGA Loom, Indy 3 and Monkey VGA) that was introduced by a recent "multi MIDI" change.
The "old Adlib" property was set too late in the iMUSE player, so the Adlib
driver would always think the game was new. I've removed this property from
the iMUSE player, since it's only function appeared to be to pass it on to
the Adlib player, and now set it in the Adlib player directly instead.

Also removed some remaining "multi MIDI" traces from the iMUSE player that
I believe were obsoleted by the aforementioned "multi MIDI" change.

svn-id: r15974