Fixed a regression where - I think - the last animation frame wasn't displayed. This was most noticeable in looping animations, in particular the spinning "Dreamer's Guild" logo in the IHNM intro.
In the process I changed the name of 'n_frames' and 'nframes' to 'maxframe'
since that's what they are. I could, of course, have added 1 to the value
instead, but that would have introduced a subtle difference compared to the
original code. Probably not a good idea at this stage.

svn-id: r16417