Added constant for the resource number for the first IHNM scene. Actually, I believe it's the wrong scene.
Currently we use resource 152, which run the scripts "CampFireSetup" and
"CampFireScene". If I use resource 151 instead it will run the scripts
"ColumnSetup" and "ColumnScene" instead. I'm speculating that this is the
"hate" speech and the rest of the intro ("ColumnScene" is a fairly long
script, as far as I can tell), but unfortunately this causes ScummVM to
crash almost immediately.

I'll need to dissect the scripts to see if I can figure out what's wrong.

svn-id: r17036