This should fix a crash which could happen when placing several objects too close to each other on the ground. (Happened to me on the first level after destroying the voodoo doll, where I'd drop the banana, the soap and the false nose close to each other on the ground after using them.)
Reasoning behind the change:

From what I understand, map_itemsMap[] contains information for each "cell"
of the map about which objects are there. Each cell can contain two objects
which are stored in the upper and lower byte of a 16-bit word.

When dropping an object, it is written into map_itemsMap[], but not just to
the indicated cell but also to a few of the surrounding ones. Presumably to
make it easier to pick it up afterwards.

When writing an object to a cell, we check if one of the bytes is already
occupied. If it is, write to the other byte. Otherwise, write to that byte.
(If both bytes are occupied, one will be overwritten.)

The old code assumed that if one byte was free at position (x,y) the same
byte would automatically be the free one in the surrounding cells. This
could cause bad values in the array, since the item was added to an
existing value, rather than replacing it.

This new code makes the check for each cell that is modified. (It also gets
rid of some code duplication.)

svn-id: r17851