My Linux ITE CD developed read errors, so Joe sent me a replacement, for which I'm very grateful.
This latest CD version turned out to be one that ScummVM didn't support. It
has - I believe - the same data files as the Linux version (with digitized
music), except for the voice file which appears to be the same as the one
from the Mac version.

In other words, it has one big-endian data file, and a couple of little-
endian ones. The technical term for this is "bloody awkward".

It should work now, though.

svn-id: r18333