Completely revisited C64 opcodes table. Matched with disassembly. Most differences in opcodes is 16bits -> 8bits operands. Instead of duplicating code I just overload getVarOrDirectWord() and fetchScriptWord(). Saves considerable amount of code.
At the moment reference is descumm. I.e. I modified the utility and later
fixed script_c64.cpp. So there could be mistakes, though I doublechecked it.
Best is to test it with real scripts.

Difference in opcodes is so big that I wonder how it worked so well before,
though we tried just few scripts.

svn-id: r18778