The IHNM intro animations run now, albeit in a slightly buggy fashion:
The intro runs a whole series of animations. Each of these use animation
slot 0, like the original. Currently this means we leak a bit of memory for
each new animation.

More noticeably, we have no way of cancelling outstanding animation events,
so unless the animations finish on their own - as some of them do - the
events will just pile up, causing the animations to run faster and faster.

I don't know if implementing the remaining cutaway opcodes will fix this
automagically, or if we need to re-think our approach...

svn-id: r18878