Add possibility to disable SDL parachute on run-time. Option --disable-sdl-parachute should be completely undocumented and not mentioned to end-users. Wrong usage may lead to system lock-up especially when used with fullscreen mode.
SDL Parachute is used to prevent system instability and do a graceful exit
if something bad happens. It catches most of signals sent to thr process.
Side effect of that that quite often when SDL application dies, core dump
isn't made on *nix systems. This feature adds possibility to disable it.

svn-id: r19146