- adds the new gui renderer also a new implementation for the classic gui  - adds a ImageMan and ImageDec class for loading and managing image files  - adds a loader for zip files which is used by the new theme and the image manager  - changes the widgets to use the new gui code  - changes the scumm dialogs to use the new gui code  - fixes a #include problem in the sky debugger with the new gui code
 To use the new gui copy gui/themes/default-theme.zip to your extrapath.
If the theme zip can not be found the gui will fallback to the classic theme.
If you want to change the gui styles use "gui_theme=classic" for the classic theme
and "gui_theme=default-theme" for the new theme.

Thanks to eriktorbjorn for testing and help with the new theme and to sev for
reviewing this patch.

svn-id: r20227