Plug in better font. Unfortunately we can't use Arial12 which Krest put on his mock-ups due to license restrictions, i.e. Microsoft prohibits any reverse engineering and modifications of it and we need to convert it into raster font.
Fortunately there are free fonts in the wild. I tried Bitstream Vera, but it
looked too wide, or crappy if horizontal resolution was altered. Next try
was Adobe Helvetica from X.org. Only requirement is to verbose copyright
notice and disclaimer.

However we will need both bold (provided) and normal wieghts of font in new
GUI for use in text inputs. This is work to do.

svn-id: r20325