If I understand things correctly, cinE used to dump the entire animDataTable element to the savefile, including data pointers. After reading the savefile, it would then test if ptr1 was NULL, to see if it should load the object.
I've extended the savefile format with a byte to indicate whether or not ptr1
was non-NULL. This seems to fix the problems I had with with loading savegames,
but of course any old savegame is now even more broken than before.

I still can't seem to get out of the room with the machine, though. Another
regression when migrating the code from cinE, or just my ability to get past
this annoying, timed puzzle?

svn-id: r21772