Tweaks to WinCE makefile.
The choice of which libraries to use (zlib, tremorce, mad, mpeg2 etc) is
now made by a single define per library towards the top of the makefile.

Given that developers tend to have different options selected on their local
builds, hopefully this will minimise the changes on aubsequent svn updates.

svn-id: r27676