Some optimisations to the DXA player to reduce the amount of needless mallocing and memcpying it does.
This commit also contains some code protected by #ifdef DXA_EXPERIMENT_MAXD
that allows for reading an extended header on a .dxa file and using a MAXD
chunk, if found, to reduce the amount of memory required to play a movie.

Whether this gets enabled or not will ultimately depend on whether the savings
are justified (i.e. if we can find a case where it makes enough difference).

svn-id: r28047