Hopefully avoiding an assert in the MSVC debug CRT library when entering an umlaut in the launcher.
It's a bad idea to cast a parameter to the "is*" functions from ctype.h to "char",
since "char" might be signed and the "is*" functions are defined to only accept
input which fits "unsigned char" or equals EOF. In this concrete example the value
of "state.ascii" is > 0x7F, thus becomes negative after the cast to "char", which
violates the parameter rules of "isprint".

Sadly it seems this is not the only place in our code, which does pass a possibly
signed parameter to an "is*" function. We might either want to change all code to
only pass an unsigned char value, change the code to do parameter validation or
fix it in another way.

svn-id: r44920