Cine: executePlayerInput: Revert r33872 as things should work without it now.
With the latest changes to executePlayerInput we
should now be able to free the girl at the bottom of the
ocean without the r33872.

The commit log from r33872 was:

"Fix for bugging moving at the bottom of the ocean when trying to free the girl from the ropes and swimming to the surface. Some global variables related to mouse position weren't being updated in executePlayerInput, now they are and things seem to work. Also enables moving in the labyrinth arcade sequence at the palace."

svn-id: r47516