Prevent "music-driver" to be set to "" on startup.
This is a regression from r50158:
"LAUNCHER: hopefully fixed music driver selection via command line".

The problem here was that the code used operator[] of HashMap
to check whether "music-driver" was specified on command line,
but that on the other hand inserted a (key, value) pair with
a default constructed Common::String as value.

svn-id: r50469