i18n: use user friendly language names in GUI
The GUI now uses the content of the Language field from the po file
header if it is present and not empty for the language selection
PopupWidget. If not present it uses the file name as before (e.g.
ru_RU).

Also update all the translation template and all the translation files.

svn-id: r51542