I18N: Fix create_translation tool when string contain special characters
The english or translated strings in the PO file might have representations of special characters (e.g. '\n') or protected characters (e.g. '\"') visible as
in the source code. However at run time ScummVM expects them to be
the special characters and not their visible representations. Now the
create_translations tool does the substitution if it finds such characters.

svn-id: r52739