SAGA: Fix an ITE intro regression
(Possibly related to bug #3076822 - ITE: Dialog repeats in intro.)

The p2_a.voc patch file wasn't played. This was because the engine
first loaded p2_a.voc as a replacement for voice resource 4. Then it
tried to load p2_a.iaf and, when it failed, removed the patch data
for the very same resource. Now it only tries to read a patch if it
hasn't already found a patch for the resource. (There may have been a
similar bug in the end credits as well.)

svn-id: r52928