TOOLS : Fixes for skycpt use, particularly on case-sensitive OS
In order to allow for work on bug #1507756 "BASS: Officer Blunt wrong animation (?)",
the skycpt tool needs to be run, but this was only previously used on case-insensitive
OS i.e. Win32. To get this to work normally on case-sensitive OS i.e. Unix, a number of
corrections were made to the hardcoded filenames to ensure the same case as the
input files.

Minor corrections :
The README was also amended to deal with the pitfall of decompressing savefiles.
Added output of each RESET.* file as processed so offending file can be seen on assertion.
Corrected segfault on missing COMPACT.TXT, so exit is more graceful.

svn-id: r52941