ANDROID: Force extra screen update when updating overlay.
This appears to work around a blank screen bug Nexus1.  I never
tracked it down, but as far as I can tell it is triggered by multiple
overlapping updates before flushing the texture to screen.  This
condition only happens in the overlay atm so an extra redraw isn't the
end of the world.

(Also remove an unused _full_screen_dirty property)

svn-id: r53801