MIDI parser: Stop hanging notes when a notes off event is called
This is the same fix as the one in r54429 for the MIDI parser of the SCI engine, based
on patch #3116608. Probably fixes (or is indirectly related to) bugs #2825928 (for
MADE), #2948281 (for SAGA), #2797283 (for tinsel), #2016549 (for SCUMM) and #2138118
(for kyra), as these engines are all using the MidiParser class internally.

svn-id: r55256