DS: Disable symbol garbage collection
Only applies when using loadable modules.
This requires a second compile run or black voodoo linker tricks.
Not implemented at this time. Disable it, so we get DS build failure
reports from buildbot (which are useful with all the DISABLE_EVERYTHING
defines).
The resulting binary might be too big to be usable.