GUI: Increment theme version.
With 50a682e7e1c69082735a07b2261d1c0c61ac12c9 we changed the resolution string
in a way which is incompatible with old resolution strings, thus all old theme
files are not usable anymore and will cause problems. To avoid that we just
increase the version string and thus it is not possible to use them anymore.